Update README.md
 Introduction

In this project, you should implement Kruskal’s algorithm, which calculates a minimum spanning tree of a given graph. You should write a program that reads a file which contains information of the graph and outputs the minimum spanning tree in the given format.

There can be many different implementation, but your implementation should have a time complexity of O(e log n), where e is the number of edges and n is the number of vertices in the given graph.

As described in the lecture slides, you can achieve O(e log n) by using min heap and disjoint set trees.

2. Requirements (Read Carefully!)

You should write a single C program, named hw3.